client/pyproject.toml
[project]
name = "epcalendar-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pillow>=11.1.0",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
]
